requirements.txt
# requirements.txt
autogen==1.0.0
python-docx==0.8.11
PyPDF2==3.0.1
markdown==3.4.3
pandas==2.1.1
openpyxl==3.1.2
python-dotenv==1.0.0
pydantic==2.4.2
fastapi==0.104.0
uvicorn==0.23.2
numpy==1.24.3
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-multipart==0.0.6
aiofiles==23.2.1
tenacity==8.2.3